+++ userspace/lib/include/sys/wait.h
@@ -5,10 +5,10 @@
 
 #define WNOHANG 1
 
-#define WIFEXITED(status) (1)
+#define WIFEXITED(status) ((status) < 128)
 #define WEXITSTATUS(status) ((status) & 0xff)
-#define WIFSIGNALED(status) (0)
-#define WTERMSIG(status) (0)
+#define WIFSIGNALED(status) ((status) >= 128)
+#define WTERMSIG(status) ((status) - 128)
 
 pid_t waitpid(pid_t pid, int *status, int options);